+++ src/main/java/com/atlauncher/network/DebugLoggingInterceptor.java
@@ -18,6 +18,7 @@
 package com.atlauncher.network;
 
 import java.io.IOException;
+import java.net.HttpURLConnection;
 
 import com.atlauncher.LogManager;
 
@@ -37,12 +38,13 @@ public final class DebugLoggingInterceptor implements Interceptor {
         Response response = chain.proceed(request);
         long t2 = System.nanoTime();
 
-        if (response.cacheResponse() != null && response.networkResponse() != null) {
-            LogManager.debug(String.format("Received cached response for %s in %.1fms", response.request().url(),
-                    (t2 - t1) / 1e6d), 3);
+        if (response.cacheResponse() != null && (response.networkResponse() == null
+                || response.networkResponse().code() == HttpURLConnection.HTTP_NOT_MODIFIED)) {
+            LogManager.debug(String.format("Received cached response code %d for %s in %.1fms", response.code(),
+                    response.request().url(), (t2 - t1) / 1e6d), 3);
         } else {
-            LogManager.debug(
-                    String.format("Received response for %s in %.1fms", response.request().url(), (t2 - t1) / 1e6d), 3);
+            LogManager.debug(String.format("Received response code %d for %s in %.1fms", response.code(),
+                    response.request().url(), (t2 - t1) / 1e6d), 3);
         }
 
         LogManager.debug(response.toString(), 5);