+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.3.5.3
 
 - Switch to using system property for Vanilla tab
+- Fix debug logging not logging if request was cached or not correctly